finance/requirements.txt
codecs50
Flask
Flask-Session
requests